Block editor: rename build-inherited-value module to resolve-styles
The builder's public entry point was renamed to resolveStyles/useResolvedStyles,
so align the filenames. Uses git mv to preserve history and keep rebases clean.

- build-inherited-value.js -> resolve-styles.js
- test/build-inherited-value.js -> test/resolve-styles.js

Updates the import in inherited-value-context.js and a doc comment in the
typography-panel test.
